restructure the project
